+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Fixed the CSS class "ib-plus" being available twice in the master CSS file.
 - Fixed full-screen mode appearing odd when used on a very small display (iPhone SE, Galaxy Fold, etc).
 - The calendar now has a minimum width of 260px.
+- Added public functions "turnOnFullScreen()" and "turnOffFullScreen()" (for toggling full-screen mode, if enabled).
 
 ### Version 0.6.2:
 - Markdown files for documentation now display correctly and types are italic.